requirement.txt
Flask
boto3
python-pptx
Pillow